refactor(calc): refactor calc commands into two files and updated autocompletion
